api/requirements.txt
fastapi[standard]
sqlmodel
psycopg2-binary
alembic
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart